Make `BytesDeserializer` public
